pyproject.toml
[project]
name = "sentiment-service"
version = "0.1.0"
description = "Backend sentiment analysis service with LangChain LCEL and FastAPI"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "typing-extensions>=4.12.2",
  "streamlit>=1.50.0",
  "requests>=2.32.5",
  "pandas>=2.3.3",
]

[tool.uv]
# nothing special; allows `uv run` and `uv sync`

[tool.pyright]
typeCheckingMode = "basic"
